Add 1h margin before sending entries to the previous day
We interpret entries with an end time in the future as entries that
happened a day before. The logic had however minute alignment issues,
and was also doing a strict comparison, which led to various entries
accidentally being sent to the wrong day. We now add a 1h margin to
reduce this risk.